tabletmanager: Split agent_rpc_actions.go into multiple files.
The monolithic file was getting unwieldy.
