+++ go/vt/tabletmanager/binlog_players_test.go
@@ -1,3 +1,7 @@
+// Copyright 2013, Google Inc. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
 package tabletmanager
 
 import (
