+++ go/vt/tabletmanager/rpc_external_reparent.go
